requirements/local.txt
-r base.txt

Werkzeug[watchdog]==3.0.1
psycopg[c]==3.1.19 
mypy==1.7.1
django-stubs[compatible-mypy]==4.2.7
pytest==7.4.4
pytest-sugar==0.9.7
djangorestframework-stubs[compatible-mypy]==3.14.5 
django-extensions==3.2.3
pytest-django==4.8.0
Faker==23.0.0